+++ src/hashing/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod common;
+pub mod multiply_shift;
+pub mod polynomial;